// src/modules/user/kyc/identity/types.ts
export interface IdentityData {
    first_name: string;
    last_name: string;
    nid?: string;
    dob: string;
    address: string;
    postcode?: string;
    staticPhone?: string;
    city?: string;
    state?: string;
    country?: string;
    metadata?: string;
    confirm?: boolean;
    metaData?: any;
}